Change backup command to giveback the error code
The past command just executed step by step, but does not
giveback the final error code. This doesn't allow the monitoring
of the backup, which could lead to a incompleted backup showed as
completed. The print "Backup completed" has been removed, as well
as the 60 sec wait at the end.

Additionally bigger MongoDBs do have a higher ressource consumption
for the dump process and get OOM killed.